CS162 Software Development/CS162 Final/StarcAI/starc-backend/requirements.txt
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
distro==1.8.0
exceptiongroup==1.2.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
Mako==1.3.0
MarkupSafe==2.1.3
nltk==3.8.1
openai==1.3.7
Pillow==10.1.0
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
pypdf==3.17.0
python-dotenv==1.0.0
regex==2023.10.3
reportlab==4.0.7
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.23
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.1.0
Werkzeug==3.0.1
zipp==3.17.0
